AQ-19204 Change all value type DTO properties to reference types
XmlSerializer throws a useless message when a property cannot be parsed into a value type because it is missing in the file. Changed the double and int types to be parsed as strings then converted to the appropriate nullable type.
With all of the properties nullable it will be possible to throw a MissingRequiredField exception for each field that is required, but missing in the file.