AQ-19245 Map Plugin CrossSectionSurvey to Framework CrossSectionSurvey
Map the CrossSection properties from the Metadata dictionary parsed from the CrossSection file.
If the specified unit does not exist in the units collection, the default unit of the GageHeight parameter will be used instead.
Channels and RelativeLocations are created if they don't already exist within the target location. If the channel name or relative location name is missing, then the default will be used (and created if needed).