AQ-19900 CrossSection - Throw if StartBank = "Unspecified"
We were detecting if the StartBank is set to an invalid value in the plug-in by seeing if the string can be mapped to the enum and then throwing if it could not.

We do this in order to produce a clearer message than we could in the business layer. However if the string equalled (with a case insensitive comparison) "Unspecified", this did not result in an exception from the plug-in.

This commit fixes that.