AQ-22008 First vertical in mean section has no segment details
Fixed bug where the framework calculates the TotalDischargePortion fails to account for the fact that the first vertical in a mean section does not have any segment details.  As a result, it was causing the framework to throw a NullReferenceException.

Also, the VerticalDepthObservation business layer DTO contains two properties - Weighting and DepthMultiplier.  These properties are used to calculate mid/mean-sections for verticals.  However, since the parser is responsible for doing these calculations, there is no need to expose these properties in our framework DTOs.