AQ-22115 Always use location identifier from file in cross-section plugin
In 17.1, the plugin would reject a location mismatch with the identifier in
the file only for a visit context. When the visit context was removed, the
behaviour was inadvertently changed to reject a mismatch in a location context
instead. Now that the framework verifies that the location of output visits
matches the context location, it is not necessary for plugins to check this at
all.